+++ src/app/pages/BuyCrypto/BuyCrypto.tsx
@@ -1,11 +1,6 @@
-import React, { FC, ReactNode, Suspense, useMemo, useState } from 'react';
+import React, { FC, useState } from 'react';
 
-import classNames from 'clsx';
-
-import Spinner from 'app/atoms/Spinner';
 import Stepper from 'app/atoms/Stepper';
-import { useAppEnv } from 'app/env';
-import ErrorBoundary from 'app/ErrorBoundary';
 import PageLayout from 'app/layouts/PageLayout';
 import ApproveStep from 'app/pages/BuyCrypto/steps/ApproveStep';
 import ExchangeStep from 'app/pages/BuyCrypto/steps/ExchangeStep';
@@ -13,101 +8,9 @@ import InitialStep from 'app/pages/BuyCrypto/steps/InitialStep';
 import { ExchangeDataInterface } from 'lib/exolix-api';
 import { T, t } from 'lib/i18n/react';
 import { useAccount, useNetwork, useStorage } from 'lib/temple/front';
-import { Redirect, Link } from 'lib/woozie';
+import { Redirect } from 'lib/woozie';
 
-import { BuyCryptoSelectors } from './BuyCypto.selectors';
-import { Cryptos } from './Cryptos';
-import { Debits } from './Debits';
-import { useTabSlug } from './hooks/use-tab-slug';
-
-interface BuyCryptoProps {
-  crypto?: boolean;
-}
-
-const BuyCrypto: FC<BuyCryptoProps> = ({ crypto = false }) => {
-  const { fullPage } = useAppEnv();
-  const tabSlug = useTabSlug();
-
-  const tabs = useMemo<
-    {
-      slug: string;
-      title: string;
-      Component: FC;
-      testID: string;
-    }[]
-  >(() => {
-    return [
-      {
-        slug: 'crypto',
-        title: t('topUpCrypto'),
-        Component: Cryptos,
-        testID: BuyCryptoSelectors.Crypto
-      },
-      {
-        slug: 'debit',
-        title: t('topUpDebit'),
-        Component: Debits,
-        testID: BuyCryptoSelectors.Debit
-      }
-    ];
-  }, []);
-
-  const { slug, Component } = useMemo(() => {
-    const tab = tabSlug ? tabs.find(currentTab => currentTab.slug === tabSlug) : null;
-    return tab ?? tabs[0];
-  }, [tabSlug, tabs]);
-
-  if (crypto) {
-    return <BuyCryptoOld />;
-  }
-  return (
-    <PageLayout
-      pageTitle={
-        <>
-          <T id="topUpBuy" />
-        </>
-      }
-    >
-      <div className="text-center my-3">
-        <T id="topUpDescription" />
-      </div>
-      <div className={classNames('-mx-4', 'shadow-top-light', fullPage && 'rounded-t-md')}>
-        <div className={classNames('w-full max-w-sm mx-auto px-10', 'flex items-center justify-center')}>
-          {tabs.map(currentTab => {
-            const active = slug === currentTab.slug;
-
-            return (
-              <Link
-                key={currentTab.slug}
-                to={lctn => ({ ...lctn, search: `?tab=${currentTab.slug}` })}
-                replace
-                className={classNames(
-                  'flex1 w-full',
-                  'text-center cursor-pointer mb-1 pb-1 pt-2',
-                  'text-gray-500 text-xs font-medium',
-                  'border-b-2',
-                  active ? 'border-primary-orange' : 'border-transparent',
-                  active ? 'text-primary-orange' : 'hover:text-primary-orange',
-                  'transition ease-in-out duration-300',
-                  'truncate'
-                )}
-                testID={currentTab.testID}
-              >
-                {currentTab.title}
-              </Link>
-            );
-          })}
-        </div>
-
-        <div className={'mx-4 mb-4 mt-6'}>
-          <SuspenseContainer whileMessage="displaying tab">{Component && <Component />}</SuspenseContainer>
-        </div>
-      </div>
-    </PageLayout>
-  );
-};
-
-const BuyCryptoOld: FC = () => (
+const BuyCrypto: FC = () => (
   <PageLayout
     pageTitle={
       <>
@@ -183,20 +86,3 @@ const BuyCryptoContent: FC = () => {
     </div>
   );
 };
-
-type SuspenseContainerProps = {
-  whileMessage: string;
-  fallback?: ReactNode;
-};
-
-const SuspenseContainer: FC<SuspenseContainerProps> = ({ whileMessage, fallback = <SpinnerSection />, children }) => (
-  <ErrorBoundary whileMessage={whileMessage}>
-    <Suspense fallback={fallback}>{children}</Suspense>
-  </ErrorBoundary>
-);
-
-const SpinnerSection: FC = () => (
-  <div className="flex justify-center my-12">
-    <Spinner theme="gray" className="w-20" />
-  </div>
-);
