--- src/app/pages/BuyCrypto/BuyCypto.selectors.ts
@@ -1,6 +1,0 @@
-export enum BuyCryptoSelectors {
-  Crypto = 'BuyCrypto/CryptoTab',
-  Debit = 'BuyCrypto/DebitTab',
-  MoonPay = 'BuyCrypto/MoonPayTab',
-  Ramp = 'BuyCrypto/RampTab'
-}
