+++ src/app/pages/SelectCrypto/SelectCrypto.selectors.ts
@@ -1,0 +1,6 @@
+export enum SelectCryptoSelectors {
+  Crypto = 'SelectCrypto/CryptoTab',
+  Debit = 'SelectCrypto/DebitTab',
+  MoonPay = 'SelectCrypto/MoonPayTab',
+  Ramp = 'SelectCrypto/RampTab'
+}
