+++ src/app/pages/SelectCrypto/assets/moonpay.png
